requirements.txt
PyYAML~=6.0
flask~=2.3.3
setuptools~=58.0.4
psutil~=5.9.5
selenium==4.8.2
pyyaml==6.0
Pillow~=10.0.0
peewee~=3.16.2
PyMySQL~=1.1.0
dataclasses-json>=0.5.2
pandas~=2.0.3
airtest==1.3.0.1
allure-pytest==2.13.2
peewee==3.16.2
pytest==7.4.0
jsonpath~=0.82.2
opencv-contrib-python~= 4.6.0.66
allure-python-commons~=2.13.1
pocoui~=1.0.89
aiohttp~=3.8.6
pytest-asyncio~= 0.21.1